Hey man,
I have tweaked a few things at Shortys:

1) Moved the subscribe section to top and deleted <p></p> between titles. Think it looks better now and that real estate is prime, but def belongs on top. Think rest of order ok.
2) changed formatting slightly of comptetion preview. used 'read more' to retain consistency and put on same line as deadline to save space. Moved date down. Think looks better (that lack of margin was bugging me)
3) Added metatags. I'm not sure if I put them in the right place. put them in the title block in the head. I assume this means they will work on all the pages, but not sure if right or not
4) I saw you added link to Black Heart at bottom of Mike's story. Nice 1
5) Changed the document background of flame_effect4 to black so hopefully should solve that white block problem
6) Amended the FAQ. Added some personal touch and potential humor but if it is too much just let me know and we can formalise it
7) Added 'who we are' story. This one talks about us. I am sorry if I missed out some stuff or misrepresented you. Please change if need be.
8) Read About and think it is fine
9) Added personal invite story for Lauren (14) and Sarah (15). Think it reads nicely and hope she will be taken by tone.

That's about it. Great to see some activity on the site.

There is a rather alarming, but not that major, bug I have noticed. It has to do with the hidden stories. If you edit a hidden story, once you save, it immideately becomes ununhidden ie: disappears from public view. I guess this is because if you edit a story from the frontend, there is no checkbox for 'hidden story' (which is right; we wouldn't want to confuse patrons), and so its status is reverted. The only worry here is, for example, if I edited Lauren's story and forgot to go to backend and check hidden story it would return an error when she tries to access. Same for about, faq and who we are. As I say, no reason to fix this as long as we are both aware of it and remember to make hidden stories hidden after each edit so people can access them.

Signed-off-by: Jeff <21stcenturynightmare@gmail.com>